devices/pi-client/requirements.txt
aiohttp==3.10.5
aiosqlite==0.20.0
pynacl==1.5.0
psutil==6.0.0
